Rename `cra-plugin-*` to `react-scripts-plugin-*` (#10)
