+++ packages/react-scripts-plugin-css-modules/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "cra-plugin-css-modules",
+  "name": "react-scripts-plugin-css-modules",
   "version": "0.2.0",
   "description": "CSS Modules plugin for Create React App",
   "main": "index.js",
@@ -7,6 +7,6 @@
   "author": "Jon Crenshaw <jdcrensh@gmail.com>",
   "license": "MIT",
   "dependencies": {
-    "cra-plugin-utils": "^0.2.0"
+    "react-scripts-plugin-utils": "^0.2.0"
   }
 }
