+++ packages/react-scripts-plugin-utils/index.js
