+++ packages/react-scripts-plugin-utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "cra-plugin-utils",
+  "name": "react-scripts-plugin-utils",
   "version": "0.2.0",
   "description": "Core utilities for Create React App plugins",
   "main": "index.js",
