+++ packages/react-scripts/template/cra.config.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = {};
